Update Main.java
The near final  command line version. Added the ability to read from an existing file list to improve time efficiency. Added a second variation of .compareVersion to allow for only one directory to be passed allowing for the boolean listExists to be removed as the logic will be handled in main.